+++ tests/tamr_client/dataset/test_dataset.py
@@ -1,18 +1,13 @@
 import pytest
-import responses
 
 import tamr_client as tc
-from tests.tamr_client import fake, utils
+from tests.tamr_client import fake
 
 
-@responses.activate
+@fake.json
 def test_from_resource_id():
     s = fake.session()
     instance = fake.instance()
-
-    dataset_json = utils.load_json("dataset.json")
-    url = tc.URL(path="datasets/1")
-    responses.add(responses.GET, str(url), json=dataset_json)
 
     dataset = tc.dataset.from_resource_id(s, instance, "1")
     assert dataset.name == "dataset 1 name"
@@ -20,13 +15,26 @@ def test_from_resource_id():
     assert dataset.key_attribute_names == ("tamr_id",)
 
 
-@responses.activate
+@fake.json
 def test_from_resource_id_dataset_not_found():
     s = fake.session()
     instance = fake.instance()
 
-    url = tc.URL(path="datasets/1")
-    responses.add(responses.GET, str(url), status=404)
-
     with pytest.raises(tc.dataset.NotFound):
         tc.dataset.from_resource_id(s, instance, "1")
+
+
+@fake.json
+def test_attributes():
+    s = fake.session()
+    dataset = fake.dataset()
+
+    attrs = tc.dataset.attributes(s, dataset)
+
+    row_num = attrs[0]
+    assert row_num.name == "RowNum"
+    assert row_num.type == tc.attribute.type.STRING
+
+    geom = attrs[1]
+    assert geom.name == "geom"
+    assert isinstance(geom.type, tc.attribute.type.Record)