+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Added function to get operation from resource ID
   - [#421](https://github.com/Datatamer/tamr-client/pull/421) Added functions for getting and replacing the transformations of a projects via `tc.transformations.get_all()` and `tc.transformations.replace_all()`
    - Added new dataclasses `Transformations` and `InputTransformations` to support these functions
+  - Moved function `tc.attribute.from_dataset_all` to `tc.dataset.attributes`
 
   **NEW FEATURES**
   - [#383](https://github.com/Datatamer/tamr-client/issues/383) Now able to create an Operation from Job resource id